Update: More Active component cleanups
