+++ src/active-trigger-factory.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import elem from './elem-factory'
 import ariasComputer from './utils/arias'
-import { useActive } from './Active/ActiveContext'
+import { useActive } from './Active/Active'
 import { useTheme } from './Theme/Theme'
 import { targetClassNames } from './utils/componentry'
